bot commands, SNR averaging

+++ src/lora_functions.h
@@ -6,7 +6,7 @@
 #include <debugconf.h>
 #include <aprs_functions.h>
 
-void OnRxDone(uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr);
+void OnRxDone(uint8_t *payload, uint16_t size, float rssi, float snr);
 void OnRxTimeout(void);
 void OnRxError(void);
 bool is_new_packet(uint8_t compBuffer[4]);
